refactor(Server): create routes.js to separate from Server.js
